refactor: :recycle: userPreview 를 user module 에서 제공하도록 수정
- close #253
- close #248
