+++ app/src/api/cursusUser/db/cursusUser.database.schema.ts
@@ -1,5 +1,6 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { HydratedDocument } from 'mongoose';
+import { user } from 'src/api/user/db/user.database.schema';
 
 export type CursusUserDocument = HydratedDocument<cursus_user>;
 
@@ -31,93 +32,6 @@ export class Cursus {
 
   @Prop()
   kind: string;
-}
-
-@Schema()
-export class Image {
-  @Prop({ required: true })
-  link?: string;
-}
-
-@Schema()
-export class UserBase {
-  @Prop({ required: true })
-  id: number;
-
-  @Prop({ required: true })
-  login: string;
-
-  @Prop()
-  url: string;
-}
-
-@Schema()
-export class User extends UserBase {
-  @Prop()
-  email: string;
-
-  @Prop()
-  firstName: string;
-
-  @Prop()
-  lastName: string;
-
-  @Prop()
-  usualFullName: string;
-
-  @Prop()
-  usualFirstName?: string;
-
-  @Prop()
-  phone: string;
-
-  @Prop({ required: true })
-  displayname: string;
-
-  @Prop()
-  kind: string;
-
-  @Prop({ required: true })
-  image: Image;
-
-  @Prop({ required: true })
-  'staff?': string;
-
-  @Prop({ required: true })
-  correctionPoint: number;
-
-  @Prop()
-  poolMonth: string;
-
-  @Prop()
-  poolYear: string;
-
-  @Prop({ required: true })
-  location?: string;
-
-  @Prop({ required: true })
-  wallet: number;
-
-  @Prop()
-  anonymizeDate: Date;
-
-  @Prop()
-  dataErasureDate: Date;
-
-  @Prop()
-  createdAt: Date;
-
-  @Prop()
-  updatedAt: Date;
-
-  @Prop()
-  alumnizedAt?: Date;
-
-  @Prop()
-  'alumni?': boolean;
-
-  @Prop()
-  'active?': boolean;
 }
 
 @Schema()
@@ -159,8 +73,8 @@ export class cursus_user {
   @Prop()
   updatedAt: Date;
 
-  @Prop({ type: User })
-  user: User;
+  @Prop()
+  user: user;
 }
 
 export const CursusUserSchema = SchemaFactory.createForClass(cursus_user);
