+++ app/src/api/scaleTeam/db/scaleTeam.database.schema.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { HydratedDocument } from 'mongoose';
-import { UserBase } from 'src/api/cursusUser/db/cursusUser.database.schema';
+import { UserBase } from 'src/api/user/db/user.database.schema';
 import { TeamBase } from 'src/api/team/db/team.database.schema';
 
 export type ScaleTeamDocument = HydratedDocument<scale_team>;
